Object orientation implemented
Redid the histograms using the object-oriented approach, treating each Bz crossing structure as an object. Also deleted some unnecessary printouts and added text wrapping to the bar chart's labels so they stop overrrunning.

Still need to generalize the histogram-generating functions in the mmsplotting module to be usable with any numerical attribute of the Structure class (need to add a dict with units information or some such)