guide field attribute, scatter plot creation
Added two new attributes to the Structure class, guide field and relative guide field (which is under construction). Added By to the timeseries plots for the sake of comparison. Plotted histograms of guide field strengths for all structures and scatter plots of guide field versus structure size for all structures.