Bartlett window filtering and resampling
Added the bartlett_interp function to the mmsmultispacecraft module to sync the timelines of the B-field data from the different spacecraft and smooth it at the same time. May want to split into two different functions at some point (in case I want to do some other sort of filtering than the Bartlett window having a separate filtering and joining function would be nice), but this will do for now.