Calculation of reciprocal vectors
Added functionality to calculate the reciprocal vectors of the MMS tetrahedron from spacecraft positional data. It was confirmed that the reciprocal vectors at each point summed to zero (or virtually zero, as they are floats). This functionality should still be checked beyond this, though(to make sure the calculation is correct).