Structure dimension determination
Added the structure_diml function in the mmsmultispacecraft module to estimate (by multiple techniques) the dimensionality of each structure, depending on the results of the MDD analysis. Created a skeleton for the STD function which will determine the velocity of the structures.

Added a REPLOT functionality to choose whether to regenerate the timeseries plots or not

Commented out modules that aren't currently being used (but may in the future)