Windowing and structure indexing improvements
Changed the section_maker function in mmsstructs to also return the indexes of the structures in each window section (so that the bartlett interpolation does not have to be done more than once). Needed to make a small change to mms_feature_search to account for the new output of section_maker.  Incorporated window-making into mms_multispacecraft_feature_search and did some general restructuring in preparation for the creation of the STD function. Plotted windows with structure indicators.