Finding average crossing times
Made two functions (datetime_avg in mmstimes and structure_crossing in mmsmultispacecraft) to aid in the calculation of the average crossing time for all spacecraft for each structure.

Also made other miscellaneous improvements, including documentation improvements, the restructuring and renaming of some of the functions in mmsstructs for better function structure, improvement of the smoothing function to allow for fixed-width smoothing.