Statistics and bug fixes
Added the capability to do multispacecraft histograms, bar charts and scatter plots of the various structure attributes, added normalized sizes and speeds to the structure attributes.

IMPORTANT: fixed an error in the elctron velocity calculation. Was using the Vix component for calculating all components of Ve, which lead to misleading/incorrect results!