Added pressures, betas and errorflags
mms_multispacecraft_feature_search now reads in the FPI ion and electron pressure tensors, converts them to total pressures, and uses this to calculate the ion beta. Also displays the FPI errorflags on the plots (as this can indicate inaccuracies in the data that may need to be corrected)

Also did back-of-the-envelope calculation of what the small and large size limits of my structure-finding algorithm are, and plotted them on the velocity-size scatter plot. An analysis of how the necessary extrema width changes the results is necessary, and should be done soon.