Electric field plotting + additional functionality
Added the ability to plot electric field data (including all of the electric field datafiles for the event). Added information about the guide field to each plot. Also attempted to do the STD method using the average MDD eigenvectors, but due to the large time variations of the eigenvectors it led to some strange results, so the function to do STD using the average MDD eigenvectors is currently unused.