j dot E calculation and statistics, accurate guide field calculation for push current sheets
Made a j_dot_e function in the plasmaparams module. Used this to find the parallel and perpendicular J dot E, and used this in statistics, both for the structures and in comparison to the total overall J dot E. Also overhauled the guide field calculation- 'core field' now refers to the minimum varying MDD direction, and is a useful quantity for plasmoids. 'guide field' is calculated using an estimate in the guide_field function in the mmsmultispacecraft module, and is useful for current sheets where the four MMS spacecraft cross the current sheet proper (e.g. push current sheets, most likely). Also changed the timeseries output plots to show j dot E instead of pressure.